Fixes #7239 - make sure the qpid client cert is deployed before the pulp migrations
Otherwise the connection to the qpid fails while running some migration file.
This doesn't happen every-time, but I've already seen puppet ordering the steps
in this way.

Also, I've realized we don't need to include the certs::pulp_parent from the
capsule.pp anymore. This was needed when we had separate modules for
capsule-certe-generate (node-certs-generate back then) and we were in the
transition phase from old installer to kafo-based one. The pulp_parent
now gets handled by the katello-intaller itself, we don't need to include that
again for capsule-certs-generate.