Fixes #25739 - don't allow statements to fail in an set -e script
When a bash script is executed with `set -e`, it will fail immediately
when a subcommand fails. From bash(1):

    -e    Exit immediately if a pipeline (which may consist of a single
          simple command), a list, or a compound command (see SHELL
          GRAMMAR above), exits with a non-zero status.

Running scripts with `-e` is generally a good idea, but this means that
we can't allow subcommands to fail anymore and then handle the error
case based on `$?`.

This commit replaces the two places where we relied on the value of `$?`
with code that will _never_ fail and at least return a dummy value which
we then can act on.