Fixes #32506: Add an nssdb_certificate type and provider
This adds a type for managing NSS database certificates (and private keys)
within an NSS database. This new type reduces the number of intermediary
entities that have to be created, for example the pkcs12 export, and
reduces the overall Puppet code required. This type aims to provide
a clean interface to managing NSS database certificates and ensure
that the NSS database is in sync with the certificates being managed.