Fixes #32511 - Add a puppet type and provider to manage an NSS database.
This commit adds and uses the new provider to manage the NSS db for the
installation. The password file location is changed from within the NSS
db directory to its parent, and the old password file will be cleaned up
when it first runs. The password itself is now generated and cached
using extlib.