Deploy managed certificate and key from build directory
This change no longer installs the RPMs that are generated when creating
certificates. Instead, the certificates that are deployed are pulled
from their locations generated on disk by the katello-ssl-tool. This
removes a layer of indirection and simplifies the creation and management
of the certificates.

With the RPMs present, a given key-pair would be created in the
ssl_build_dir, then deployed to /etc/pki/katello-certs-tools and then
copied and managed somewhere else on disk, often /etc/pki/katello.