Include entire directory of certificates in the created tarball
The current tarball method only includes the generated RPMs which
are copied to a host and then those RPMs are installed placing
the certificates into /etc/pki/katello-certs-tools. This change
includes the certificates directly alongside the RPM which will
include them on a host at /root/ssl-build/<hostname>. This is an
additive step that will allow future changes to deploy the certificates
directly from the build directory rather than the RPMs.