Add a certificate_file type
This shows adding a new type for managing certificate files copied from
a source location. This type creats a File type under the hood for
managing the file and its attributes. This introduces a server_cert
parameter that allows adding additional validation of the source
certificate into the provider on top of the validations provided by
the underlying File resource.

The source_cert has to be implemented as a property and validated
inside the provider code. The validate block of a type executes during
catalog compilation and does not give access to the certificate on disk.