Fixes #32886: Allow passing a client certificate to use for Artemis communication
Introduce a new parameter artemis_client_cert that allows a user
to pass in a certificate to use for communication with Artemis. This
certificate is used to calculate the client DN and placed in the
truststore. If this new parameter is not supplied, the previous behavior
of generating a keypair and deploying it is used. The previous behavior
remains the default.

If a keypair had been generated and deployed previously, and a
client certificate is supplied, the previously generated and deployed
keypair is removed from the system.