Refs #32585: Allow producing certificate subject with spaces between elements
Adds an option to produce a certificate subject with spaces between
elements. This is necessary if a certificate has been generated in
standard format rather than RFC2253 format. One specific use case is
when Artemis does strict checking of the certificate subject with
the subject in the /etc/tomcat/cert-users.properties file. They
need to be identical.