Deploy Apache certificates from build directory
This introduces a new wrapper object key_pair for deploying and
managing a certificate and private key based on input files. This change
skips the creation and installation of RPMs, instead opting to deploy
the key and certificate directly from the build directory. This skips
a lot of overhead that comes with the RPMs.