Add new type to build a certificate and key bundle
Adds a new type to create a bundle of a certificate and private key
that relies on certificate and private key paths being supplied. The
current key_bundle type is tied closely to the underlying katello-certs-tools
based types. This new type is more widely usable across use cases and
combines the ability to manage file properties.