Correctly pass the path to the tempfile when adding keystore certificates
This was previously pass the name of the File object in Ruby to
keytool creating a temporary file at the location that the module
is executed at to serve as the temporary file. This means that
the result continues to work -- a keystore certificates is imported
into the final destination but it was leaving behind residual files.